feat: improved and completed inventory
Now you can take and save an item with you, drop it if you dont want it, or directly replace it with another one. The text updates with the name of the item. Logic improved and tested to avoid errors. Now you can't have an item and select/move other items at the same time.