refactor: player health system integrated with persecution
Integrates the player health system with the persecution manager, enabling damage interactions with enemies.

Moved agentFollowPos to Vampire folder and renamed to FollowPlayerAgent.

Moved VidaJugador to Player folder and renamed to PersecutionManager, extracted logic which dealed with player health and player health UI to corresponding classes in Player/Health folder